Caclulations with floating point numbers

+++ FractalPart.pde
@@ -1,8 +1,8 @@
 class FractalPart {
-  public double fraction;
+  public float fraction;
   public Author author;
   
-  public FractalPart(double fraction, Author author) {
+  public FractalPart(float fraction, Author author) {
     this.fraction = fraction;
     this.author = author;
   }
